requirements.txt
flask==2.3.3
flask-cors==3.0.10
PyPDF2==3.0.1
python-docx==0.8.11
chardet==4.0.0
python-dotenv==0.19.0
flask-session==0.4.0
docusign-esign==3.20.0
pyjwt==2.3.0
fpdf==1.7.2
pdfplumber==0.10.3
SQLAlchemy==1.4.23
openai==1.3.7
Werkzeug==2.3.7
nltk==3.8.1
numpy>=1.24.3
httpx==0.24.1